requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.2
colorama==0.4.4
comtypes==1.1.14
distlib==0.3.4
filelock==3.6.0
Flask==2.1.1
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
multidict==6.0.4
numpy==1.24.2
openai==0.27.4
platformdirs==2.5.1
pypiwin32==223
python-dotenv==1.0.0
pyttsx3==2.90
pywin32==306
requests==2.28.2
six==1.16.0
tqdm==4.65.0
urllib3==1.26.15
virtualenv==20.14.0
Werkzeug==2.1.0
yarl==1.8.2
